requirements.txt
langchain_community
streamlit
langchain
langchain_groq 
langchain_core
python_dotenv
python-pptx
Flask